requirements.txt
pytest==5.3.5
diskcache==4.1.0
python-slugify==4.0.0